feat:interfaces/config.tsの作成
